gtk3: Fix headerbar transparency on some apps
Some applications draw additional headerbar elements on top of the
main csd headerbar. This results in the headerbar appearing
non-transparent, even if transparency is enabled, as two (or more)
backgrounds with 0.97 alpha are stacked on top of each other.

Fix this by making the background transparent on every headerbar
that's descendant of another headerbar.