gtk3: Fix padding for .text-button.image-button
The combined horizontal padding for .text-button.image-button labels
was 13px (5px for the button + 8px for the label), while the padding
for plain .text-button is 12px.

This resulted in labels being jumpy at least on nautilus 3.32
path-bar when switching directories, and possibly elsewhere.

Fix this by reducing label paddings by 1px. Also add default 2px
paddings to the label for when it is in between two images, and
therefore the :first-child, :last-child or :only-child selectors
wouldn't apply.