gnome-shell: Replace menu SVG assets with CSS
This works around the issue with border-images being scaled
incorrectly when fractional scaling is enabled, is less prone to odd
visual glitches, and overall simpler implementation.

This reimplemention has the usual issue, where the borders don't
appear transparent, as they do in the previous implementation. This
results from the background color fill correctly extending below the
borders. It could be corrected using the 'background-clip' CSS
property, which unfortunately remains unsupported by gnome-shell
version 3.32.

Additionally there are some minor differences in the appearance of
borders and shadows.