gnome-shell: Replace modal.svg asset with CSS
This works around the issue with border-images being scaled
incorrectly when fractional scaling is enabled, is less prone to odd
visual glitches, and overall simpler implementation.

The reimplementation should be near pixel-perfect apart from the
missing shadow, as the box-shadow is not displayed due to some
gnome-shell painting oddities. There are also some minor differences
in the appearance of the round corners.