gnome-shell: Don't use SVG assets on notifications
Replace message SVG assets in notification messages, and menu.svg
asset in notification banners with CSS.

This works around the issue with border-images being scaled
incorrectly for banners when fractional scaling is enabled, is less
prone to odd visual glitches, and overall simpler implementation.

Unfortunately, due to current gnome-shell limitations, the appearance
of notifications on the light variant can't be replicated perfectly.
Only possibility would be using border-images, which in turn would
result in scaling issues. Updates to the handling of background
and/or border related CSS properties in gnome-shell could possibly
change this in the future.