gtk3: Add padding to xfce4-panel buttons
Attempt to fix zero horizontal padding in many xfce4-panel 4.14
buttons, by adding it to labels and icons that are next to labels.

For some unknown reason the padding doesn't seem to apply on some
icons (at least the tasklist button icons). Also the padding can't
be applied directly to the button element, again for unknown reason.