build: Move Plank theme from extra to common
The extra directory was originally meant for themes that couldn't be
installed with the build system, but required manual installation by
the user (i.e. Chrome theme, and later Plank).

As Plank theme is installed with automake to $XDG_DATA_DIRS/themes
like rest of the themes, from where it can be enabled with the 'Gtk+'
theme option [1], it only makes sense to move it to common.

[1] https://github.com/horst3180/arc-theme/issues/570