build: Always define GNOME_SHELL_3_32_OR_NEWER
When building the gnome-shell theme, the build system needs to know
if it is built for shell version 3.32 or later, since the Arc-lighter
theme variant does not exist for earlier versions.

The GNOME_SHELL_3_32_OR_NEWER conditional definition was invoked only
when building the shell theme, which resulted in build failure if
the build of gnome-shell theme was disabled. Fix this by always
defining the conditional.

This is a bit of an messy workaround, but should work until either
the build system is rewritten with better solution, or the lighter
variant is ported for all supported gnome-shell versions.

Fixes https://github.com/jnsh/arc-theme/issues/24