build: Make optipng optional
optipng is not required for building the theme, but only used to
optimize the rendered PNG assets, which will reduce the size of the
built theme marginally. Make it optional, and only used if available
on the build environment, to avoid dependency hell with some
distributions, e.g. Clear Linux [1].

[1] https://github.com/jnsh/arc-theme/issues/52#issuecomment-642966596