Remove openbox theme
The openbox theme was included from the arc-design repo under the
impression that it was specifically created for that project.
However, it was actually copied from an existing project [1].

It doesn't make sense to have the theme developed in two separate
locations. The themes supported in this repo should be limited to
GTK and GTK-based desktops, as it was with the original Arc-theme
project.

The Arc openbox theme remains available from its original repository.

[1] https://github.com/dglava/arc-openbox