gtk3: Update nautilus path-bar styling
Nautilus 3.38 removed the special width-maximized style class from
the path-bar, and now always uses the maximized design. Update the
path-bar styling for Arc accordingly.

As an unwanted side effect, this also changes the path-bar style for
Nautilus 3.30 to 3.36 in similar fashion, but there doesn't seem to
be any way to work around this.

Closes https://github.com/jnsh/arc-theme/issues/85