gnome-shell: Draw panel-button backgrounds with box-shadow
The button background extends behind the border, which caused their
labels to appear not vertically centered in overview, where the
border is transparent. This can't be corrected with background-clip
CSS property since it is not supported in gnome-shell.

Work around this by drawing the background with inset box-shadow,
which doesn't extend to the border-box. This is what the current
default shell theme does as well, although not necessarily for the
same reason.