gnome-shell: Drop Arc-Lighter variant
Arc-Lighter GNOME Shell theme attempted to be a complete light
recoloring to the original Arc Shell themes. This is much unlike
the Arc-Lighter GTK themes, that only changed the dark sidebar color.

The lighter shell theme was never on par in quality with the other
variants, and the light design had some problems due to upstream
theme being inteded to be dark, especially with the new overview in
GNOME Shell 40.

For these reasons, drop the sub-par Arc-Lighter GNOME Shell theme
starting from version 40.