gtk3: Apply padding only to selected xfce4-panel buttons
The padding is necessary as spacing between the icon and the label in
application menu and tasklist buttons, and also as general padding
for the button especially around the label.

However it's not necessary when the button only has single icon.
There was attempt to prevent this by not applying the padding to
image:only-child, but that didn't work always as expected.

Apply the padding only to specific buttons, as needed, from now on.

Fixes: https://github.com/jnsh/arc-theme/issues/162