xfwm4: Increase resize border width
The resize area for windows in xfwm4 only consists of the actual,
border area, and not any area outside the decorations like with many
other window managers [1]. Due to this, thin decoration borders
cause windows resizing to be difficult.

Increase border width on Arc decorations from 2px to 4px, which
should considerably improve the situation, while still keeping
the borders thin enough to fit the Arc design.

This also fixes the appearance of the outer border of dark window
decorations, where the titlebar has 1px dark border, when it is 2px
wide around rest of the window.

[1] https://gitlab.xfce.org/xfce/xfwm4/-/issues/176

Fixes: https://github.com/jnsh/arc-theme/issues/63