gtk3: Add background color to nautilus path-bar
Add button background color to the nautilus path-bar-box container,
for identical appearance with regular path-bars and stack-switchers.

The problem is that this by itself breaks the appearance of hovered
buttons inside the path-bar-box, since there are two stacked
transparent button backgrounds. Work around this by using transparent
white backgrounds for hovered buttons, with the transparency tuned
to make them look as close as possible with regular path-bar buttons.

Closes: https://github.com/jnsh/arc-theme/issues/96