gtk3, gtk4: Install gtk-dark.css for Arc-Dark
With the new "dark style preference" introduced with GNOME 42, when
dark style is requested while using a theme that doesn't have a
gtk-dark.css, GTK will use the default dark theme Adwaita-Dark,
instead of falling back to just using the non-dark version of the
selected theme, which was the old behavior when applications
requested dark theme. Fix this situation in Arc-Dark by adding the
gtk-dark.css files just like with other variants.

Fixes: https://github.com/jnsh/arc-theme/issues/180